+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+- Change default compression to zstd
+- Add option to set compression level, zstd defaults to 15
 - Change default temporal property to datetime
 - Enable Converter.columns list and tuple types
 - Add BaseConverter get_columns hook to customize columns after reading the file